+++ lib/grape/exceptions/invalid_response.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Grape
   module Exceptions
     class InvalidResponse < Base