+++ lib/grape/util/json.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Grape
   if Object.const_defined? :MultiJson
     Json = ::MultiJson