+++ lib/grape/util/reverse_stackable_values.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require_relative 'stackable_values'
 
 module Grape