+++ lib/grape/validations/types/dry_type_coercer.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'dry-types'
 
 module DryTypes