+++ spec/grape/middleware/formatter_spec.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'spec_helper'
 
 describe Grape::Middleware::Formatter do
@@ -41,7 +43,7 @@ describe Grape::Middleware::Formatter do
     end
 
     context 'xml' do
-      let(:body) { 'string' }
+      let(:body) { +'string' }
       it 'calls #to_xml if the content type is xml' do
         body.instance_eval do
           def to_xml